+++ packages/node/src/configure/configure.module.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// // Copyright 2020-2025 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
 import { DynamicModule, Global, Module } from '@nestjs/common';
@@ -25,7 +25,12 @@ export class ConfigureModule {
   }
   static async register(): Promise<DynamicModule> {
     const { nodeConfig, project } = await ConfigureModule.getInstance();
-
+    return this.registerManual(nodeConfig, project);
+  }
+  static registerManual(
+    nodeConfig: NodeConfig,
+    project: SubqueryProject,
+  ): DynamicModule {
     return {
       module: ConfigureModule,
       providers: [