+++ packages/common-cosmos/CHANGELOG.md
@@ -6,6 +6,47 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.4.5] - 2025-07-22
+### Fixed
+- Templates not being included in the published package (#339)
+
+## [5.4.4] - 2025-07-17
+### Changed
+- Update @subql/common (#337)
+
+## [5.4.3] - 2025-07-01
+### Changed
+- Update `@subql/common` (#332)
+
+## [5.4.2] - 2025-05-01
+### Changed
+- Re release of previously failed release
+
+## [5.4.1] - 2025-05-01
+### Changed
+- Update `@subql/common`
+
+## [5.4.0] - 2025-04-24
+### Changed
+- Update `@subql/common` (#322)
+
+## [5.3.0] - 2025-03-10
+### Changed
+- Update `@subql/common` (#318)
+
+## [5.2.0] - 2025-02-10
+### Changed
+- Update `@subql/types-cosmos`
+
+## [5.1.2] - 2025-02-04
+### Changed
+- Update header year to 2025 (#307)
+- Update `@subql/common`
+
+## [5.1.1] - 2025-01-28
+### Changed
+- Update `@subql/common` dependency
+
 ## [5.1.0] - 2024-12-17
 ### Changed
 - Update `@subql/common` dependency
@@ -174,7 +215,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix chainTypes not being in deployments
 
 ## [0.0.5] - 2022-06-15
-[Unreleased]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.1.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.4.5...HEAD
+[5.4.5]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.4.4...common-cosmos/5.4.5
+[5.4.4]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.4.3...common-cosmos/5.4.4
+[5.4.3]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.4.2...common-cosmos/5.4.3
+[5.4.2]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.4.1...common-cosmos/5.4.2
+[5.4.1]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.4.0...common-cosmos/5.4.1
+[5.4.0]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.3.0...common-cosmos/5.4.0
+[5.3.0]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.2.0...common-cosmos/5.3.0
+[5.2.0]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.1.2...common-cosmos/5.2.0
+[5.1.2]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.1.1...common-cosmos/5.1.2
+[5.1.1]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.1.0...common-cosmos/5.1.1
 [5.1.0]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.0.6...common-cosmos/5.1.0
 [5.0.6]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.0.5...common-cosmos/5.0.6
 [5.0.5]: https://github.com/subquery/subql-cosmos/compare/common-cosmos/5.0.4...common-cosmos/5.0.5