+++ packages/types/CHANGELOG.md
@@ -6,6 +6,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.1.3] - 2025-07-17
+### Changed
+- Update \`@subql/types-core\` (#337)
+
+## [4.1.2] - 2025-07-01
+### Changed
+- Update `@subql/types-core` (#332)
+
+## [4.1.1] - 2025-04-24
+### Changed
+- Update `@subql/types-core` (#322)
+
+## [4.1.0] - 2025-02-10
+### Changed
+- Use forked version of `@cosmjs/stargate` (#310)
+
+## [4.0.1] - 2025-02-04
+### Changed
+- Update block type to allow access to decoded transactions, messages and events (#305)
+- Update header year to 2025 (#307)
+- Update `@subql/types-core`
+
 ## [4.0.0] - 2024-10-23
 ### Changed
 - Bump `@subql/types-core` dependency (#292)
@@ -136,7 +158,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix chainTypes not being in deployments
 
 ## [0.0.5] - 2022-06-15
-[Unreleased]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.0.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.1.3...HEAD
+[4.1.3]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.1.2...types-cosmos/4.1.3
+[4.1.2]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.1.1...types-cosmos/4.1.2
+[4.1.1]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.1.0...types-cosmos/4.1.1
+[4.1.0]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.0.1...types-cosmos/4.1.0
+[4.0.1]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/4.0.0...types-cosmos/4.0.1
 [4.0.0]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/3.5.3...types-cosmos/4.0.0
 [3.5.3]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/3.5.2...types-cosmos/3.5.3
 [3.5.2]: https://github.com/subquery/subql-cosmos/compare/types-cosmos/3.5.1...types-cosmos/3.5.2